lambda/api_wrapper/requirements.txt
awslabs.mcp-lambda-handler==0.1.4
requests==2.32.4